+++ src/app/api-server/src/http_server.rs
@@ -129,6 +129,10 @@ pub async fn build_server(
         "/system/metrics",
         axum::routing::get(observability::metrics::metrics_handler),
     )
+    .route(
+        "/system/info",
+        axum::routing::get(observability::build_info::build_info_handler),
+    )
     .merge(kamu_adapter_http::openapi::router().into())
     .fallback(unknown_fallback_handler)
     .layer(axum::extract::Extension(gql_schema))